# governance_migration_bundle/apps/governance-api/src/main.py
"""
BioDiscoveryAI Governance API

Purpose:
  Stable API layer for governance operations.
  Tools should use this API instead of directly mutating governance tables.

Run:
  export DATABASE_URL="postgresql://user:pass@localhost:5432/biodiscovery"
  uvicorn src.main:app --reload
"""

from __future__ import annotations

import os
from dataclasses import dataclass
from functools import lru_cache
from typing import Any, Callable
from uuid import UUID

import jwt
import psycopg
from psycopg.rows import dict_row
from psycopg.types.json import Jsonb
from fastapi import Depends, FastAPI, HTTPException, status
from fastapi.security import HTTPAuthorizationCredentials, HTTPBearer
from jwt import PyJWKClient
from pydantic import BaseModel, Field


def database_url() -> str | None:
    return os.getenv("DATABASE_URL") or os.getenv("BDAI_DATABASE_URL")


bearer_scheme = HTTPBearer(auto_error=False)


@dataclass(frozen=True)
class Principal:
    subject: str
    issuer: str
    audience: str
    roles: frozenset[str]
    scopes: frozenset[str]


def required_auth_setting(name: str) -> str:
    value = os.getenv(name)
    if not value:
        raise HTTPException(
            status_code=status.HTTP_503_SERVICE_UNAVAILABLE,
            detail=f"{name} is not configured",
        )
    return value


@lru_cache(maxsize=4)
def jwks_client(jwks_uri: str) -> PyJWKClient:
    return PyJWKClient(jwks_uri)


def decode_token(token: str) -> dict[str, Any]:
    issuer = required_auth_setting("GOVERNANCE_JWT_ISSUER")
    audience = required_auth_setting("GOVERNANCE_JWT_AUDIENCE")
    hs256_secret = os.getenv("GOVERNANCE_JWT_HS256_SECRET")
    jwks_uri = os.getenv("GOVERNANCE_OIDC_JWKS_URI")

    try:
        if hs256_secret:
            return jwt.decode(
                token,
                hs256_secret,
                algorithms=["HS256"],
                issuer=issuer,
                audience=audience,
            )
        if jwks_uri:
            signing_key = jwks_client(jwks_uri).get_signing_key_from_jwt(token)
            return jwt.decode(
                token,
                signing_key.key,
                algorithms=["RS256", "ES256"],
                issuer=issuer,
                audience=audience,
            )
        raise HTTPException(
            status_code=status.HTTP_503_SERVICE_UNAVAILABLE,
            detail="GOVERNANCE_OIDC_JWKS_URI or GOVERNANCE_JWT_HS256_SECRET is not configured",
        )
    except jwt.PyJWTError as exc:
        raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED, detail="Invalid bearer token") from exc


def claim_set(claims: dict[str, Any], name: str) -> frozenset[str]:
    value = claims.get(name, [])
    if isinstance(value, str):
        return frozenset(part for part in value.split() if part)
    if isinstance(value, list):
        return frozenset(str(part) for part in value)
    return frozenset()


def current_principal(credentials: HTTPAuthorizationCredentials | None = Depends(bearer_scheme)) -> Principal:
    if credentials is None or credentials.scheme.lower() != "bearer":
        raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED, detail="Bearer token is required")

    claims = decode_token(credentials.credentials)
    subject = str(claims.get("sub") or "")
    if not subject:
        raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED, detail="Token subject is required")

    roles = claim_set(claims, "roles")
    realm_access = claims.get("realm_access")
    if isinstance(realm_access, dict):
        roles = roles | claim_set(realm_access, "roles")

    return Principal(
        subject=subject,
        issuer=str(claims.get("iss")),
        audience=os.getenv("GOVERNANCE_JWT_AUDIENCE", ""),
        roles=roles,
        scopes=claim_set(claims, "scope") | claim_set(claims, "scp"),
    )


def require_permission(
    allowed_roles: set[str],
    allowed_scopes: set[str],
) -> Callable[[Principal], Principal]:
    def dependency(principal: Principal = Depends(current_principal)) -> Principal:
        if principal.roles & allowed_roles or principal.scopes & allowed_scopes:
            return principal
        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN, detail="Insufficient governance permission")

    return dependency


require_read = require_permission(
    {"governance_admin", "governance_reader", "governance_auditor"},
    {"governance:read"},
)
require_package_write = require_permission(
    {"governance_admin", "package_writer"},
    {"packages:write"},
)
require_evidence_write = require_permission(
    {"governance_admin", "evidence_writer"},
    {"evidence:write"},
)


app = FastAPI(
    title="BioDiscoveryAI Governance API",
    version="1.0.0",
    description="Governed API for lineage, deployability, packages, images, evidence, and release readiness.",
)


class PackageCreate(BaseModel):
    package_manager: str = Field(min_length=1, max_length=64)
    package_name: str = Field(min_length=1, max_length=256)
    package_version: str = Field(min_length=1, max_length=128)
    package_description: str = Field(min_length=1, max_length=2000)
    inclusion_reason: str = Field(min_length=1, max_length=2000)
    functional_role: str = Field(min_length=1, max_length=256)
    security_risk_level: str = Field(pattern="^(low|medium|high|critical)$")
    license_name: str | None = Field(default=None, max_length=256)
    package_source_url: str | None = Field(default=None, max_length=2048)
    package_sha256: str | None = Field(default=None, pattern="^[A-Fa-f0-9]{64}$")


class EvidenceCreate(BaseModel):
    image_id: UUID
    build_id: UUID | None = None
    evidence_type: str = Field(min_length=1, max_length=128)
    evidence_tool: str = Field(min_length=1, max_length=128)
    evidence_uri: str = Field(min_length=1, max_length=2048)
    evidence_sha256: str = Field(pattern="^[A-Fa-f0-9]{64}$")
    evidence_summary: dict[str, Any] = Field(default_factory=dict, max_length=100)
    pass_fail_status: str = Field(pattern="^(pass|fail|warning|accepted_risk)$")
    generated_by: str = Field(min_length=1, max_length=256)


def conn(role: str | None = None, principal: Principal | None = None):
    DATABASE_URL = database_url()
    if not DATABASE_URL:
        raise HTTPException(status_code=500, detail="DATABASE_URL/BDAI_DATABASE_URL is not configured")
    connection = psycopg.connect(DATABASE_URL, connect_timeout=5)
    if role or principal:
        with connection.cursor() as cur:
            if role:
                cur.execute(f"SET ROLE {role}")
            if principal:
                cur.execute("SELECT set_config('app.actor', %s, false)", (principal.subject,))
                cur.execute("SELECT set_config('app.roles', %s, false)", (",".join(sorted(principal.roles)),))
                cur.execute("SELECT set_config('app.scopes', %s, false)", (",".join(sorted(principal.scopes)),))
    return connection


@app.get("/health")
def health() -> dict[str, str]:
    return {"status": "ok"}


@app.get("/ready")
def ready() -> dict[str, str]:
    with conn() as c:
        with c.cursor() as cur:
            cur.execute("SELECT 1")
            cur.fetchone()
    return {"status": "ready"}


@app.get("/images/{image_id}/deployable")
def image_deployable(image_id: UUID, principal: Principal = Depends(require_read)) -> dict[str, Any]:
    with conn(role="bdai_readonly", principal=principal) as c:
        with c.cursor() as cur:
            cur.execute("SELECT container_governance.is_image_deployable(%s)", (str(image_id),))
            result = cur.fetchone()
    return {"image_id": str(image_id), "deployable": bool(result[0]) if result else False}


@app.get("/images/{image_id}/release-certified")
def image_release_certified(image_id: UUID, principal: Principal = Depends(require_read)) -> dict[str, Any]:
    with conn(role="bdai_readonly", principal=principal) as c:
        with c.cursor() as cur:
            cur.execute("SELECT container_governance.is_release_certified(%s)", (str(image_id),))
            result = cur.fetchone()
    return {"image_id": str(image_id), "release_certified": bool(result[0]) if result else False}


@app.get("/lineage/container/{image_id}")
def container_lineage(image_id: UUID, principal: Principal = Depends(require_read)) -> dict[str, Any]:
    with conn(role="bdai_readonly", principal=principal) as c:
        with c.cursor(row_factory=dict_row) as cur:
            cur.execute(
                """
                SELECT *
                FROM container_governance.v_container_full_lineage
                WHERE image_id = %s
                """,
                (str(image_id),),
            )
            rows = cur.fetchall()
    return {"image_id": str(image_id), "lineage": rows}


@app.post("/packages")
def create_package(
    payload: PackageCreate,
    principal: Principal = Depends(require_package_write),
) -> dict[str, Any]:
    with conn(role="bdai_app_runtime", principal=principal) as c:
        with c.cursor(row_factory=dict_row) as cur:
            cur.execute(
                """
                INSERT INTO container_governance.approved_packages (
                    package_manager,
                    package_name,
                    package_version,
                    package_description,
                    inclusion_reason,
                    functional_role,
                    security_risk_level,
                    license_name,
                    package_source_url,
                    package_sha256
                )
                VALUES (%s,%s,%s,%s,%s,%s,%s,%s,%s,%s)
                RETURNING package_id
                """,
                (
                    payload.package_manager,
                    payload.package_name,
                    payload.package_version,
                    payload.package_description,
                    payload.inclusion_reason,
                    payload.functional_role,
                    payload.security_risk_level,
                    payload.license_name,
                    payload.package_source_url,
                    payload.package_sha256,
                ),
            )
            row = cur.fetchone()
            c.commit()
    return {"package_id": str(row["package_id"])}


@app.post("/evidence")
def register_evidence(
    payload: EvidenceCreate,
    principal: Principal = Depends(require_evidence_write),
) -> dict[str, Any]:
    with conn(role="bdai_app_runtime", principal=principal) as c:
        with c.cursor(row_factory=dict_row) as cur:
            cur.execute(
                """
                INSERT INTO container_governance.container_evidence (
                    image_id,
                    build_id,
                    evidence_type,
                    evidence_tool,
                    evidence_uri,
                    evidence_sha256,
                    evidence_summary,
                    pass_fail_status,
                    generated_by
                )
                VALUES (%s,%s,%s,%s,%s,%s,%s::jsonb,%s,%s)
                RETURNING evidence_id
                """,
                (
                    str(payload.image_id),
                    str(payload.build_id) if payload.build_id else None,
                    payload.evidence_type,
                    payload.evidence_tool,
                    payload.evidence_uri,
                    payload.evidence_sha256,
                    Jsonb(payload.evidence_summary),
                    payload.pass_fail_status,
                    payload.generated_by,
                ),
            )
            row = cur.fetchone()
            c.commit()
    return {"evidence_id": str(row["evidence_id"])}