# governance_migration_bundle/apps/governance-api/tests/snapshots/openapi.json
{
  "components": {
    "schemas": {
      "EvidenceCreate": {
        "properties": {
          "build_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Build Id"
          },
          "evidence_sha256": {
            "pattern": "^[A-Fa-f0-9]{64}$",
            "title": "Evidence Sha256",
            "type": "string"
          },
          "evidence_summary": {
            "maxProperties": 100,
            "title": "Evidence Summary",
            "type": "object"
          },
          "evidence_tool": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Evidence Tool",
            "type": "string"
          },
          "evidence_type": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Evidence Type",
            "type": "string"
          },
          "evidence_uri": {
            "maxLength": 2048,
            "minLength": 1,
            "title": "Evidence Uri",
            "type": "string"
          },
          "generated_by": {
            "maxLength": 256,
            "minLength": 1,
            "title": "Generated By",
            "type": "string"
          },
          "image_id": {
            "format": "uuid",
            "title": "Image Id",
            "type": "string"
          },
          "pass_fail_status": {
            "pattern": "^(pass|fail|warning|accepted_risk)$",
            "title": "Pass Fail Status",
            "type": "string"
          }
        },
        "required": [
          "image_id",
          "evidence_type",
          "evidence_tool",
          "evidence_uri",
          "evidence_sha256",
          "pass_fail_status",
          "generated_by"
        ],
        "title": "EvidenceCreate",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "PackageCreate": {
        "properties": {
          "functional_role": {
            "maxLength": 256,
            "minLength": 1,
            "title": "Functional Role",
            "type": "string"
          },
          "inclusion_reason": {
            "maxLength": 2000,
            "minLength": 1,
            "title": "Inclusion Reason",
            "type": "string"
          },
          "license_name": {
            "anyOf": [
              {
                "maxLength": 256,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "License Name"
          },
          "package_description": {
            "maxLength": 2000,
            "minLength": 1,
            "title": "Package Description",
            "type": "string"
          },
          "package_manager": {
            "maxLength": 64,
            "minLength": 1,
            "title": "Package Manager",
            "type": "string"
          },
          "package_name": {
            "maxLength": 256,
            "minLength": 1,
            "title": "Package Name",
            "type": "string"
          },
          "package_sha256": {
            "anyOf": [
              {
                "pattern": "^[A-Fa-f0-9]{64}$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Package Sha256"
          },
          "package_source_url": {
            "anyOf": [
              {
                "maxLength": 2048,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Package Source Url"
          },
          "package_version": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Package Version",
            "type": "string"
          },
          "security_risk_level": {
            "pattern": "^(low|medium|high|critical)$",
            "title": "Security Risk Level",
            "type": "string"
          }
        },
        "required": [
          "package_manager",
          "package_name",
          "package_version",
          "package_description",
          "inclusion_reason",
          "functional_role",
          "security_risk_level"
        ],
        "title": "PackageCreate",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Governed API for lineage, deployability, packages, images, evidence, and release readiness.",
    "title": "BioDiscoveryAI Governance API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/evidence": {
      "post": {
        "operationId": "register_evidence_evidence_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EvidenceCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Register Evidence Evidence Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Register Evidence"
      }
    },
    "/health": {
      "get": {
        "operationId": "health_health_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Health Health Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Health"
      }
    },
    "/images/{image_id}/deployable": {
      "get": {
        "operationId": "image_deployable_images__image_id__deployable_get",
        "parameters": [
          {
            "in": "path",
            "name": "image_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Image Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Image Deployable Images  Image Id  Deployable Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Image Deployable"
      }
    },
    "/images/{image_id}/release-certified": {
      "get": {
        "operationId": "image_release_certified_images__image_id__release_certified_get",
        "parameters": [
          {
            "in": "path",
            "name": "image_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Image Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Image Release Certified Images  Image Id  Release Certified Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Image Release Certified"
      }
    },
    "/lineage/container/{image_id}": {
      "get": {
        "operationId": "container_lineage_lineage_container__image_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "image_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Image Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Container Lineage Lineage Container  Image Id  Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Container Lineage"
      }
    },
    "/packages": {
      "post": {
        "operationId": "create_package_packages_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PackageCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Create Package Packages Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Package"
      }
    },
    "/ready": {
      "get": {
        "operationId": "ready_ready_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Ready Ready Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Ready"
      }
    }
  }
}